+++ MDM/App-install/App browser-security Auto-install.sh
@@ -31,6 +31,7 @@ installomatorOptions="BLOCKING_PROCESS_ACTION=tell_user_then_quit" # Separated b
 # Fill the variable "item" above with a label.
 # Script will run this label through Installomator.
 ######################################################################
+# v.  9.2.3 : Only kill the caffeinate process we create
 # v.  9.2.2 : A bit more logging on succes.
 # v.  9.2.1 : Better logging handling and installomatorOptions fix.
 ######################################################################
@@ -55,7 +56,6 @@ fi
 caffeinatepid=$!
 caffexit () {
     kill "$caffeinatepid"
-    pkill caffeinate
     exit $1
 }