+++ MDM/install swiftDialog direct.sh
@@ -16,7 +16,8 @@
 #  https://github.com/Installomator/Installomator
 #
 ######################################################################
-scriptVersion="9.6"
+scriptVersion="9.7"
+# v.  9.7   : 2022-12-19 : Only kill the caffeinate process we create
 # v.  9.6   : 2022-11-15 : GitHub API call is first, only try alternative if that fails.
 # v.  9.5   : 2022-09-21 : change of GitHub download
 # v.  9.4   : 2022-09-14 : downloadURL can fall back on GitHub API.
@@ -53,7 +54,6 @@ fi
 caffeinatepid=$!
 caffexit () {
     kill "$caffeinatepid" || true
-    pkill caffeinate || true
     printlog "[LOG-END] Status $1"
     exit $1
 }