+++ fragments/labels/teamviewer.sh
@@ -1,7 +1,8 @@
 teamviewer)
     name="TeamViewer"
     type="pkgInDmg"
-    packageID="com.teamviewer.teamviewer"
+    # packageID="com.teamviewer.teamviewer"
+    versionKey="CFBundleShortVersionString"
     pkgName="Install TeamViewer.app/Contents/Resources/Install TeamViewer.pkg"
     downloadURL="https://download.teamviewer.com/download/TeamViewer.dmg"
     appNewVersion=$(curl -fs "https://www.teamviewer.com/en/download/mac-os/" | grep "Current version" | cut -d " " -f3 | cut -d "<" -f1)